+++ src/main/java/com/socialising/services/model/Hashtag.java
@@ -2,7 +2,7 @@ package com.socialising.services.model;
 
 public class Hashtag {
 
-    private Integer hashtagId;
+    private Long hashtagId;
 
     private String hashtag;
 }
